time_t 64bit fixes for relayd and relayctl: - fix statistics - set INT_MAX limit on session timeouts - make sure we dont use to large session timeouts in pf redirects and   openssl
tested with old and new time_t

ok florian@

+++ usr.sbin/relayd/relayd.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: relayd.h,v 1.166 2013/03/09 14:43:06 bluhm Exp $	*/
+/*	$OpenBSD: relayd.h,v 1.167 2013/03/10 23:32:53 reyk Exp $	*/
 
 /*
  * Copyright (c) 2006 - 2012 Reyk Floeter <reyk@openbsd.org>
@@ -230,7 +230,7 @@ struct ctl_stats {
 	objid_t			 id;
 	int			 proc;
 
-	u_int			 interval;
+	u_int64_t		 interval;
 	u_int64_t		 cnt;
 	u_int32_t		 tick;
 	u_int32_t		 avg;
